fix: add zod parser
