+++ solutions/swb-app/src/validatorHelper.ts
@@ -5,7 +5,6 @@
 
 import Boom from '@hapi/boom';
 import { ValidatorResult } from 'jsonschema';
-import { ZodTypeAny } from 'zod';
 
 function processValidatorResult(validatorResult: ValidatorResult): void {
   if (!validatorResult.valid) {
@@ -22,22 +21,4 @@ function processValidatorResult(validatorResult: ValidatorResult): void {
   }
 }
 
-function validateAndParse<T>(parser: ZodTypeAny, data: unknown): T {
-  const parsed = parser.safeParse(data);
-
-  if (parsed.success) {
-    return parsed.data;
-  }
-
-  const errorMessages = parsed.error.issues.map((issue) => {
-    return `${issue.path.join('.')}: ${issue.message}`;
-  });
-
-  throw Boom.badRequest(
-    errorMessages.reduce((fullMessage, message) => {
-      return `${fullMessage}. ${message}`;
-    })
-  );
-}
-
-export { processValidatorResult, validateAndParse };
+export { processValidatorResult };
